chore(engine): Phase 2 debt closeout — ammoTypeShift, formations.js, startup guard, dirIndex (#286 #291 #292 #293)
- Remove tautological firerSPs >= threshold check from ammoTypeShift(); domain-expert confirmed
  unreachable (LOB §5.6 — threshold is derived from the same SP tier firerSPs falls into)
- Extract FORMATION_EFFECTS, ACTIVITY_EFFECTS, formationEffect(), activityEffect() from weapons.js
  into formations.js; move tests to formations.test.js
- Add _loadedMapPaths/_loadedScenarioPaths Set guards to loadMap()/loadScenario() that warn on
  repeated default-path calls (enforcement for M4+ game-loop misuse)
- Add dirIndex 0–5 range validation to hexEntryCost(); throws RangeError on invalid input;
  add 4 tests covering dirIndex 6, -1, NaN, and valid 0–5
